refactor(attempts): share command result handling
