+++ lib/hive/commands/stage_action.rb
@@ -8,7 +8,7 @@ require "hive/task_resolver"
 require "hive/workflows"
 require "hive/task_action"
 require "hive/attempts/context"
-require "hive/attempts/entrypoint"
+require "hive/attempts/command_dispatch"
 require "hive/conditions/transition_guard"
 
 module Hive
@@ -31,6 +31,7 @@ module Hive
     # `hive-stage-action` envelope is emitted at the end.
     class StageAction
       include Hive::Schemas::EnvelopeEmitter
+      include Hive::Attempts::CommandDispatch
 
       def initialize(verb, target, project: nil, from: nil, json: false,
                      recover_merged_error_reason: nil, durable: false,
@@ -65,39 +66,8 @@ module Hive
 
       private
 
-      def dispatch_durable
-        task = resolve_task
-        intended_stage = Hive::Workflows.for_verb(@verb).fetch(:target)
-        result = (@attempt_entrypoint || Hive::Attempts::Entrypoint.new).dispatch(
-          task: task,
-          intended_stage: intended_stage,
-          argv: durable_worker_argv(task),
-          interactive: true
-        )
-        handle_durable_failure!(result) unless result.exit_status.zero?
-
-        result
-      end
-
-      def handle_durable_failure!(result)
-        if result.status == :lost
-          exit(result.exit_status) if @json && result.stdout_emitted?
-
-          raise Hive::ConcurrentRunError,
-                "durable attempt lost before producing a receipt for #{@target}: #{result.attempt_id}"
-        end
-
-        if @json && !result.stdout_emitted?
-          raise Hive::AttemptExecutionError.new(
-            "durable attempt #{result.attempt_id} finished with #{result.outcome || 'an error'} " \
-            "(exit #{result.exit_status}) before emitting JSON",
-            exit_code: result.exit_status,
-            attempt_id: result.attempt_id,
-            outcome: result.outcome
-          )
-        end
-
-        exit(result.exit_status)
+      def durable_intended_stage(_task)
+        Hive::Workflows.for_verb(@verb).fetch(:target)
       end
 
       def durable_worker_argv(task)
