+++ wiki/testing.md
@@ -65,6 +65,7 @@ task default: :test
 | `markers_test.rb` | `Hive::Markers` — set/get round-trip, attribute quoting, last-marker semantics. |
 | `atomic_file_test.rb` | `Hive::AtomicFile` — atomic file replacement plus the shared directory-fsync helper's real flush and unsupported-platform (`NotImplementedError` / `EINVAL` / `ENOTSUP` / `EBADF`) behavior. |
 | `attempts/*_test.rb`, `daemon/attempt_loss_healer_test.rb` | Durable task attempts — record/schema/CAS edges, competing claim/expiry, artifact-plus-dependency-verdict generation duplicates/successors, capacity, detached session lifecycle, framed log/client replay, PID-reuse-safe adoption and orphan cleanup, marker/git evidence precedence, legacy backfill, mutation-free dirty capture, and restart-persistent bounded loss healing. |
+| `commands/{run,stage_action}_test.rb`, `integration/{run_error_envelope,run_stage_action}_test.rb` | Durable command-result interpretation — run/stage-specific admitted argv and intended stages plus the shared `CommandDispatch` policy for success receipts, lost attempts, non-zero receipts with or without worker JSON, exact exit propagation, and one command-shaped error document when a failed worker emitted none. |
 | `task_journal_test.rb`, `conditions/*_test.rb`, `task_projection*_test.rb`, `task_action_conditions_test.rb` | Generation-scoped conditions — strict journal durability/attempt integrity including short-write completion and partial-append rollback, authenticated replay with future-schema/forged-attempt rejection, registry/evidence lifecycle, generation/HEAD reconciliation, deterministic supersession, telemetry-tail-stable snapshot validation/replay, gate/migration/shadow modes, execute boundary ordering, canonical TaskAction/status consumption, and sanitized task-1849 replay across missing/corrupt/stale snapshots with durable-attempt metadata and live-observation sentinels. |
 | `lock_test.rb` | `Hive::Lock` — acquire/release, stale-PID detection, commit lock parallelism. |
 | `worktree_test.rb` | `Hive::Worktree` — create attach-vs-new, dependency override stacking (incl. narrow-refspec and origin-ahead-of-local **and** local-ahead-of-origin placeholders), explicit remote-head selection over a same-named tag, stalled-transport deadline/process-group cleanup, empty placeholder re-pointing, fail-closed preservation when the emptiness check errors, local-only prerequisite fallback, real-commit preservation, PR-head materialization/retry/failure handling, delete-failure errors, `local_branch_ref_exists?` blank-name guard, remove, exists?, pointer round-trip, prefix validation. |
